in progress...

+++ src/main/java/avadamedia/kinocms/repository/UserRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends CrudRepository<User, Long> {
     @Query(value = "SELECT MAX(id) FROM users", nativeQuery = true)
-    Integer getMaxId();
+    Long getMaxId();
 }
