+++ src/main/java/avadamedia/kinocms/service/impl/UserServiceImpl.java
@@ -44,7 +44,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Integer getMaxId() {
+    public Long getMaxId() {
         return repository.getMaxId();
     }
 }
